requirements.txt
torch
torchvision
matplotlib==3.4.2
numpy==1.21.2
pandas==1.3.0
seaborn==0.11.1
pillow==8.3.1
opencv-python
scipy
tqdm==4.67.1